+++ bin/easiest_localization.dart
@@ -1,4 +1,5 @@
-import 'package:easiest_localization/easiest_localization.dart';
+import 'package:easiest_localization/src/gen/generator.dart';
+import 'package:easiest_localization/src/gen/generator_config.dart';
 import 'package:easiest_localization/src/loader/config_loader.dart';
 import 'package:easiest_localization/src/loader/language_localization.dart';
 
@@ -10,9 +11,9 @@ void main(List<String> args) {
   }
   final Generator generator = Generator(config);
   final (String, List<LanguageLocalization>) result = generator.generate();
-  final double end = (DateTime.now().millisecondsSinceEpoch - start) / 1000;
+  final int end = (DateTime.now().millisecondsSinceEpoch - start);
   // ignore: avoid_print
   print(
-    '[EASIEST_LOCALIZATION] Localizations generation completed in $end seconds with [${result.$2.map((LanguageLocalization it) => it.name).join(', ')}] locales',
+    '[EASIEST_LOCALIZATION] Localizations generation completed in ${end}ms with [${result.$2.map((LanguageLocalization it) => it.name).join(', ')}] locales',
   );
 }