+++ example/android/app/src/main/kotlin/dev/alphamikle/example/MainActivity.kt
@@ -1,6 +1,0 @@
-package dev.alphamikle.example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}