+++ lib/easiest_localization.dart
@@ -1,4 +1,3 @@
 library easiest_localization;
 
-export './src/gen/generator.dart';
-export './src/gen/generator_config.dart';
+export 'src/provider/localization_provider.dart';