+++ lib/src/provider/localization_provider.dart
@@ -1,0 +1,9 @@
+abstract interface class LocalizationProvider<L, M> {
+  String get name;
+
+  List<L> get supportedLocales;
+
+  Future<M> fetchLocalization(L locale);
+
+  bool canLoad(L locale);
+}