+++ lib/src/template/pubspec_template.dart
@@ -1,8 +1,10 @@
 import '../gen/generator_config.dart';
 
 String assetsPubspecTemplate(GeneratorConfig config) => '''
+# Generated with easiest_localization: https://pub.dev/packages/easiest_localization
 name: ${config.packageName}
 description: ${config.packageDescription}
+publish_to: none
 version: ${config.packageVersion}
 
 environment:
@@ -13,4 +15,5 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+  easiest_localization: ${config.easiestLocalizationVersion}
 ''';